sync works

+++ tools/commands/publish.js
@@ -10,17 +10,19 @@ export async function publish(...flags){
         const target = new Target(pkg.dir, flags, []);
         if (target.packageJson.private)
             continue;
-        await fs.rename(
-            join(target.rootDir, './package.json'),
-            join(target.rootDir, './.package.json')
-        );
-        const content = await target.getPublishPackageJson();
+        if (target.tsConfig) {
+            await fs.rename(
+                join(target.rootDir, './package.json'),
+                join(target.rootDir, './.package.json')
+            );
+            const content = await target.getPublishPackageJson();
 
-        await fs.writeFile(
-            join(target.rootDir, 'package.json'),
-            content,
-            'utf-8'
-        )
+            await fs.writeFile(
+                join(target.rootDir, 'package.json'),
+                content,
+                'utf-8'
+            )
+        }
         const cp = await spawn("corepack", "yarn npm publish".split(' '), {
             cwd: target.rootDir,
             stdio: "pipe",
@@ -37,10 +39,12 @@ export async function publish(...flags){
         if (!error) {
             target.log(`published`);
         }
-        await fs.rm(join(target.rootDir, './package.json'));
-        await fs.rename(
-            join(target.rootDir, './.package.json'),
-            join(target.rootDir, './package.json')
-        );
+        if (target.tsConfig) {
+            await fs.rm(join(target.rootDir, './package.json'));
+            await fs.rename(
+                join(target.rootDir, './.package.json'),
+                join(target.rootDir, './package.json')
+            );
+        }
     }
 }
